requirements.txt
alembic==1.8.0
#factory==1.2##Error Occured
factory_boy==3.2.1
fastapi[all]==0.79.0
hashids==1.3.1
jose==1.0.0
passlib==1.7.4
pydantic==1.10.6
pytest==7.1.2
python_jose==3.3.0
requests
SQLAlchemy==1.4.37
SQLAlchemy_Utils==0.38.2
typing_extensions==4.3.0
yarg==0.1.9
mysqlclient
python-dotenv
ulid-py
pytest-cov
python-multipart
boto3
pillow
bcrypt
PyJWT[crypt]
cryptography